+++ src/main/java/com/cucumber/utils/basicstepdefs/sql/SqlSteps.java
@@ -8,6 +8,7 @@ import cucumber.api.java.en.Then;
 import cucumber.runtime.java.guice.ScenarioScoped;
 
 import java.io.IOException;
+import java.sql.SQLException;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -22,56 +23,97 @@ public class SqlSteps {
     @Given("SQL data source from file path \"{cstring}\"")
     public void setDataSource(String filePath) throws IOException {
         this.dataSource = ResourceUtils.readProps(filePath);
-        SqlClient.Builder builder = new SqlClient.Builder();
-        builder.driver(dataSource.getProperty("driver").trim())
-                .user(dataSource.getProperty("username")).pwd(dataSource.getProperty("password"))
-                .url(dataSource.getProperty("url"));
-        this.client = builder.build();
+        this.client = new SqlClient(dataSource.getProperty("url"), dataSource.getProperty("username"),
+                dataSource.getProperty("password"), dataSource.getProperty("driver").trim());
     }
 
     @Then("SQL execute query \"{cstring}\"")
-    public void executeQuery(String query) {
-        this.client.executeQuery(query);
+    public void executeQuery(String query) throws SQLException {
+        try {
+            this.client.connect();
+            this.client.prepareStatement(query);
+            this.client.executeQuery();
+        } finally {
+            this.client.close();
+        }
     }
 
     @Then("SQL execute query \"{cstring}\" and compare result with")
-    public void executeQueryAndCompare(String query, List expected) {
-        this.result = client.executeQuery(query);
-        Cucumbers.compare(expected, result, false, true);
+    public void executeQueryAndCompare(String query, List expected) throws SQLException {
+        try {
+            this.client.connect();
+            this.client.prepareStatement(query);
+            this.result = client.executeQueryAndGetRsAsList();
+            Cucumbers.compare(expected, result, false, true);
+        } finally {
+            this.client.close();
+        }
     }
 
     @Then("SQL execute query \"{cstring}\" and poll for {int}s while comparing result with")
-    public void executeQueryAndPollAndCompare(String query, int pollDuration, List expected) {
-        Cucumbers.pollAndCompare(expected, () -> client.executeQuery(query), false, true);
+    public void executeQueryAndPollAndCompare(String query, int pollDuration, List expected) throws SQLException {
+        try {
+            this.client.connect();
+            this.client.prepareStatement(query);
+            Cucumbers.pollAndCompare(expected, () -> client.executeQueryAndGetRsAsList(), false, true);
+        } finally {
+            this.client.close();
+        }
     }
 
     @Then("SQL execute update \"{cstring}\"")
-    public void executeUpdate(String sql) {
-        this.client.executeUpdate(sql);
+    public void executeUpdate(String sql) throws SQLException {
+        try {
+            this.client.connect();
+            this.client.prepareStatement(sql);
+            this.client.executeUpdate();
+        } finally {
+            this.client.close();
+        }
     }
 
     @Then("SQL INSERT into table \"{cstring}\" the following data")
-    public void insertDataInsideTable(String table, List data) {
-        String sql = "INSERT INTO " + table + " (%s) values (%s)";
-        ((List<Map<String, String>>) data).forEach(map -> {
-            StringBuilder columns = new StringBuilder();
-            StringBuilder values = new StringBuilder();
-            map.forEach((k, v) -> {
-                columns.append(columns.length() != 0 ? "," : "").append(k);
-                values.append(values.length() != 0 ? "," : "").append(v.equals("null") ? null : "'" + v + "'");
+    public void insertDataInsideTable(String table, List data) throws SQLException {
+        try {
+            this.client.connect();
+            String sql = "INSERT INTO " + table + " (%s) values (%s)";
+            ((List<Map<String, String>>) data).forEach(map -> {
+                StringBuilder columns = new StringBuilder();
+                StringBuilder values = new StringBuilder();
+                map.forEach((k, v) -> {
+                    columns.append(columns.length() != 0 ? "," : "").append(k);
+                    values.append(values.length() != 0 ? "," : "").append(v.equals("null") ? null : "'" + v + "'");
+                });
+                try {
+                    this.client.prepareStatement(String.format(sql, columns.toString(), values.toString()));
+                    this.client.executeUpdate();
+                } catch (SQLException e) {
+                    throw new RuntimeException(e);
+                }
             });
-            this.client.executeUpdate(String.format(sql, columns.toString(), values.toString()));
-        });
+        } finally {
+            this.client.close();
+        }
     }
 
     @Then("SQL UPDATE table \"{cstring}\" WHERE \"{cstring}\" with the following data")
-    public void updateDataFromTable(String table, String cond, List data) {
-        String sql = "UPDATE " + table + " SET %s WHERE " + cond;
-        StringBuilder assignmentValues = new StringBuilder();
-        ((List<Map<String, String>>) data).forEach(map -> {
-            map.forEach((k, v) -> assignmentValues.append(assignmentValues.length() != 0 ? "," : "")
-                    .append(k).append("=").append(v.equals("null") ? null : "'" + v + "'"));
-            this.client.executeUpdate(String.format(sql, assignmentValues.toString()));
-        });
+    public void updateDataFromTable(String table, String cond, List data) throws SQLException {
+        try {
+            this.client.connect();
+            String sql = "UPDATE " + table + " SET %s WHERE " + cond;
+            StringBuilder assignmentValues = new StringBuilder();
+            ((List<Map<String, String>>) data).forEach(map -> {
+                map.forEach((k, v) -> assignmentValues.append(assignmentValues.length() != 0 ? "," : "")
+                        .append(k).append("=").append(v.equals("null") ? null : "'" + v + "'"));
+                try {
+                    this.client.prepareStatement(String.format(sql, assignmentValues.toString()));
+                    this.client.executeUpdate();
+                } catch (SQLException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        } finally {
+            this.client.connect();
+        }
     }
 }