+++ src/main/java/com/cucumber/utils/clients/http/HttpResponseLoggerInterceptor.java
@@ -1,7 +1,6 @@
 package com.cucumber.utils.clients.http;
 
 import org.apache.http.HttpEntity;
-import org.apache.http.HttpException;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpResponseInterceptor;
 import org.apache.http.entity.StringEntity;
@@ -11,7 +10,6 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
-import java.io.UnsupportedEncodingException;
 import java.util.Arrays;
 
 public class HttpResponseLoggerInterceptor implements HttpResponseInterceptor {
@@ -19,13 +17,16 @@ public class HttpResponseLoggerInterceptor implements HttpResponseInterceptor {
     private Logger log = LogManager.getLogger();
 
     @Override
-    public void process(HttpResponse response, HttpContext context) throws HttpException, IOException {
+    public void process(HttpResponse response, HttpContext context) {
         log.debug("--- HTTP RESPONSE ---");
         log.debug("Response STATUS: {}", response.getStatusLine());
         log.debug("Response HEADERS: {}", Arrays.asList(response.getAllHeaders()));
         log.debug("Response BODY:{}{}", () -> System.lineSeparator(), () -> {
-            HttpEntity entity = response.getEntity();
             if (response == null) {
+                return null;
+            }
+            HttpEntity entity = response.getEntity();
+            if (entity == null) {
                 return null;
             }
             String content = null;
@@ -36,10 +37,11 @@ public class HttpResponseLoggerInterceptor implements HttpResponseInterceptor {
                 return "Cannot consume HTTP response: " + e.getMessage();
             } finally {
                 try {
-                    if (response != null && content != null) {
+                    EntityUtils.consume(entity);
+                    if (content != null) {
                         response.setEntity(new StringEntity(content));
                     }
-                } catch (UnsupportedEncodingException e) {
+                } catch (IOException e) {
                     log.error(e);
                 }
             }